Rename Okta lib function and file name
Jira ticket: CAMS-401
